Desenvolvendo a classe de modelo e o carregador de .obj

+++ libs/vec3f.h
@@ -3,15 +3,15 @@
 
 typedef struct vec3f
 {
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 } vec3f_t;
 
 typedef struct vec2f
 {
-    float x;
-    float y;
+    double x;
+    double y;
 }vec2f_t;
 
 #endif
